changed: visualization interface and visualizations to use float data instead of S16
git-svn-id: https://xbmc.svn.sourceforge.net/svnroot/xbmc/branches/AudioEngine@32817 568bbfeb-2a22-0410-94d2-cc84cf5bfa90

+++ xbmc/visualizations/DirectXSpectrum/directx_spectrum.cpp
@@ -239,7 +239,7 @@ extern "C" void Start(int iChannels, int iSamplesPerSec, int iBitsPerSample, con
   z_angle = 0.0f;
 }
 
-extern "C" void AudioData(const short* pAudioData, int iAudioDataLength, float *pFreqData, int iFreqDataLength)
+extern "C" void AudioData(const float* pAudioData, int iAudioDataLength, float *pFreqData, int iFreqDataLength)
 {
   int i,c;
   int y=0;
